+++ src/main/java/com/springboot/bankapplication/dto/AccountDto.java
@@ -2,10 +2,12 @@ package com.springboot.bankapplication.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 public record AccountDto(
          Long id,
          String holderName,
-         double balance
+         BigDecimal balance
 ) {
 
 }
